requirements.txt
groq
numpy
requests
python-dotenv